+++ packages/patrol_cli/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Handle uninstallation of the app after tests finish when using AGP 8.2+. (#2535)
 - Add `--coverage-packages` flag to `patrol test` that specifies which packages should be included 
 in the coverage report. (#2536)
+- Fix skipping tests with coverage collection enabled. (#2539)
 
 ## 3.5.0